+++ CHANGELOG.rst
@@ -66,6 +66,11 @@ Changed
   convert those to scorers using :func:`sklearn.metrics.make_scorer`. See :pr:`2814` by
   :user:`auguste-probabl`.
 
+- **Breaking change:** :meth:`EstimatorReport.diagnose` and :meth:`EstimatorReport.add_checks`
+  have been moved to the newly added :meth:`EstimatorReport.checks` accessor and have been
+  renamed to :meth:`EstimatorReport.checks.summarize` and :meth:`EstimatorReport.checks.add`
+  respectively. See :pr:`2833` by :user:`GaetandeCast`.
+
 Added
 -----